fix(inference): cast BF16 tensors to float32 before .numpy() in _tiled_reconstruct
Under CUDA AMP (bfloat16 autocast), model outputs are BF16 tensors.
NumPy does not support BFloat16, so calling .numpy() directly raised
TypeError and silently failed for every volume in misfit_inspect.

Add .float() before .cpu().numpy() on both reconstruction and mask
outputs. Add a regression test that passes BF16 tensors through
_tiled_reconstruct to ensure this is caught going forward.

+++ misfit/inference/inference_runners.py
@@ -79,7 +79,7 @@ def _tiled_reconstruct(
                 )
                 with torch.no_grad(), amp_ctx:
                     output = model_fn(tensor)
-                recon = output["reconstruction"].squeeze().cpu().numpy()
+                recon = output["reconstruction"].squeeze().float().cpu().numpy()
                 if denorm_patches:
                     # The model was trained to predict per-patch-normalised values
                     # (zero mean, unit variance per patch cube). Undo that using the
@@ -90,7 +90,7 @@ def _tiled_reconstruct(
                     recon = recon * patch_std + patch_mean
                 recon_out[di:di + pd_, hi:hi + ph_, wi:wi + pw_] = recon
                 mask_out[di:di + pd_, hi:hi + ph_, wi:wi + pw_] = (
-                    output["mask"].squeeze().cpu().numpy()
+                    output["mask"].squeeze().float().cpu().numpy()
                 )
 
     return recon_out, mask_out
